make training publicable
